refactor: CreateTripView, TripDetailView를 TripPlanView로 통합
- TripPlanVie에서 isEditMode로 수정, 완료 페이지를 변경할 수 있게 했습니다.
