+++ src/router/index.ts
@@ -1,10 +1,4 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '@/views/HomeView.vue'
-import TripView from '@/views/TripView.vue'
-import CreateTripView from '@/views/CreateTripView.vue'
-import SearchView from '@/views/SearchView.vue'
-import LogView from '@/views/LogView.vue'
-import TripDetailView from '@/views/TripDetailView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -12,33 +6,32 @@ const router = createRouter({
     {
       path: '/',
       name: 'home',
-      component: HomeView,
+      component: () => import('@/views/HomeView.vue'),
     },
     {
       path: '/trips',
       name: 'trips',
-      component: TripView,
+      component: () => import('@/views/TripView.vue'),
     },
     {
       path: '/create-trip',
       name: 'create-trip',
-      component: CreateTripView,
+      component: () => import('@/views/TripPlanView.vue'),
     },
     {
       path: '/search',
       name: 'search',
-      component: SearchView,
+      component: () => import('@/views/SearchView.vue'),
     },
     {
-      path: '/trip/:id',
+      path: '/trips/:id',
       name: 'trip-detail',
-      component: TripDetailView,
-      props: true, // URL 파라미터를 props로 전달
+      component: () => import('@/views/TripPlanView.vue'),
     },
     {
       path: '/log',
       name: 'log',
-      component: LogView,
+      component: () => import('@/views/LogView.vue'),
     },
   ],
 })
